+++ packages/expo-modules-core/ios/ModuleRegistryAdapter/EXModuleRegistryAdapter.h
@@ -14,11 +14,11 @@ NS_SWIFT_NAME(ModuleRegistryAdapter)
 @property (nonnull, nonatomic, readonly) EXModuleRegistryProvider *moduleRegistryProvider;
 
 - (instancetype)initWithModuleRegistryProvider:(nonnull EXModuleRegistryProvider *)moduleRegistryProvider
-__deprecated_msg("Expo modules are now being automatically registered. You can delete it altogether.");
+__deprecated_msg("Expo modules are now automatically registered. You can remove this method call.");
 
 - (nonnull NSArray<id<RCTBridgeModule>> *)extraModulesForModuleRegistry:(nonnull EXModuleRegistry *)moduleRegistry;
 
 - (nonnull NSArray<id<RCTBridgeModule>> *)extraModulesForBridge:(nonnull RCTBridge *)bridge
-__deprecated_msg("Expo modules are now being automatically registered. You can replace it with an empty array.");
+__deprecated_msg("Expo modules are now automatically registered. You can replace this with an empty array.");
 
 @end